+++ packages/react-behave/CHANGELOG.md
@@ -22,6 +22,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - `props.render` has been renamed to `props.children` for components `ClickOutside`, `Dropdown` and `MergeRefs`.
 - Remove `renderMethod`.
 
+### Enhancements
+
+- Create `ResizeObserver` component.
+
 ## 0.3.0 (July 5, 2018)
 
 ### Breaking changes